scripts/ni_review/pyproject.toml
[project]
name = "news-intelligence"
version = "8.0.0"
description = "AI-powered news aggregation and analysis with local LLM inference"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core Framework
    "fastapi>=0.104.0,<1.0",
    "pydantic>=2.5.0",
    "uvicorn[standard]>=0.24.0",
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # HTTP & Requests
    "requests>=2.31.0",
    "httptools>=0.6.0",
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.25.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    # ML / Embeddings (Ollama for LLM; sentence-transformers for embeddings)
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "nltk>=3.8.0",
    # Web scraping & parsing
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
    "lxml>=4.9.0",
    "chardet>=5.0.0",
    # PDF document processing (document_processing_service: pdfplumber → pymupdf → pdfminer)
    "pdfplumber>=0.11.0",
    "pdfminer.six>=20221105",
    "pymupdf>=1.23.0",
    # Scheduling & monitoring
    "schedule>=1.2.0",
    "psutil>=5.9.0",
    "prometheus-client>=0.17.0",
    # Security
    "pyjwt>=2.12.0",
    "cryptography>=41.0.0",
    "bcrypt>=4.1.0",
    # Config & utilities
    "python-dotenv>=1.0.0",
    "gunicorn>=21.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "pyyaml>=5.1",
    "click>=8.1.0",
]

[project.optional-dependencies]
finance = ["chromadb", "sentence-transformers"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.pytest.ini_options]
pythonpath = ["api"]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "requires_api: needs a running API (e.g. localhost:8001); skipped in CI by default",
    "requires_db: needs PostgreSQL; may skip if unavailable",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "api/_archived",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
# E501: line length — formatter handles wrapping
# E402: imports after sys.path/bootstrap (legacy modules)
# E722: bare except (legacy; fix incrementally)
# F401/F403: unused imports / star-imports (__init__ barrels) — tighten later
# N803/N806/N818: ML conventions (X, y) and local ALL_CAPS config blocks in functions
# UP035: typing.Dict/List deprecation — optional modernization
ignore = ["E501", "E402", "E722", "F401", "F403", "N803", "N806", "N818", "UP035"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
